extract from zk
